+++ src/muya/lib/ui/quickInsert/config.js
@@ -1,24 +1,24 @@
-import paragraphIcon from '../../assets/icons/paragraph.svg'
-import htmlIcon from '../../assets/icons/html.svg'
-import hrIcon from '../../assets/icons/horizontal_line.svg'
-import frontMatterIcon from '../../assets/icons/front_matter.svg'
-import header1Icon from '../../assets/icons/header_1.svg'
-import header2Icon from '../../assets/icons/header_2.svg'
-import header3Icon from '../../assets/icons/header_3.svg'
-import header4Icon from '../../assets/icons/header_4.svg'
-import header5Icon from '../../assets/icons/header_5.svg'
-import header6Icon from '../../assets/icons/header_6.svg'
-import newTableIcon from '../../assets/icons/new_table.svg'
-import bulletListIcon from '../../assets/icons/bullet_list.svg'
-import codeIcon from '../../assets/icons/code.svg'
-import quoteIcon from '../../assets/icons/quote.svg'
-import todoListIcon from '../../assets/icons/todolist.svg'
-import mathblockIcon from '../../assets/icons/math.svg'
-import orderListIcon from '../../assets/icons/order_list.svg'
-// import flowchartIcon from '../../assets/icons/flowchart.svg'
-// import sequenceIcon from '../../assets/icons/sequence.svg'
-// import mermaidIcon from '../../assets/icons/mermaid.svg'
-// import vegaIcon from '../../assets/icons/chart.svg'
+import paragraphIcon from '../../assets/pngicon/paragraph/2.png'
+import htmlIcon from '../../assets/pngicon/html/2.png'
+import hrIcon from '../../assets/pngicon/horizontal_line/2.png'
+import frontMatterIcon from '../../assets/pngicon/front_matter/2.png'
+import header1Icon from '../../assets/pngicon/heading_1/2.png'
+import header2Icon from '../../assets/pngicon/heading_2/2.png'
+import header3Icon from '../../assets/pngicon/heading_3/2.png'
+import header4Icon from '../../assets/pngicon/heading_4/2.png'
+import header5Icon from '../../assets/pngicon/heading_5/2.png'
+import header6Icon from '../../assets/pngicon/heading_6/2.png'
+import newTableIcon from '../../assets/pngicon/new_table/2.png'
+import bulletListIcon from '../../assets/pngicon/bullet_list/2.png'
+import codeIcon from '../../assets/pngicon/code/2.png'
+import quoteIcon from '../../assets/pngicon/quote_block/2.png'
+import todoListIcon from '../../assets/pngicon/todolist/2.png'
+import mathblockIcon from '../../assets/pngicon/math/2.png'
+import orderListIcon from '../../assets/pngicon/order_list/2.png'
+import flowchartIcon from '../../assets/pngicon/flowchart/2.png'
+import sequenceIcon from '../../assets/pngicon/sequence/2.png'
+import mermaidIcon from '../../assets/pngicon/mermaid/2.png'
+import vegaIcon from '../../assets/pngicon/chart/2.png'
 import { isOsx } from '../../config'
 
 const COMMAND_KEY = isOsx ? '⌘' : '⌃'
@@ -40,27 +40,6 @@ const COMMAND_KEY = isOsx ? '⌘' : '⌃'
 
 
 export const quicInsertObj = {
-  // 'diagram': [{
-  //   title: 'Vega Chart',
-  //   subTitle: 'Render flow chart by vega-lite.js.',
-  //   label: 'vega-lite',
-  //   icon: vegaIcon
-  // }, {
-  //   title: 'Flow Chart',
-  //   subTitle: 'Render flow chart by flowchart.js.',
-  //   label: 'flowchart',
-  //   icon: flowchartIcon
-  // }, {
-  //   title: 'Sequence Diagram',
-  //   subTitle: 'Render sequence diagram by js-sequence.',
-  //   label: 'sequence',
-  //   icon: sequenceIcon
-  // }, {
-  //   title: 'Mermaid',
-  //   subTitle: 'Render Diagram by mermaid.',
-  //   label: 'mermaid',
-  //   icon: mermaidIcon
-  // }],
   'basic block': [{
     title: 'Paragraph',
     subTitle: 'Lorem Ipsum is simply dummy text',
@@ -166,5 +145,26 @@ export const quicInsertObj = {
     label: 'ul-task',
     shortCut: `⌥${COMMAND_KEY}X`,
     icon: todoListIcon
+  }],
+  'diagram': [{
+    title: 'Vega Chart',
+    subTitle: 'Render flow chart by vega-lite.js.',
+    label: 'vega-lite',
+    icon: vegaIcon
+  }, {
+    title: 'Flow Chart',
+    subTitle: 'Render flow chart by flowchart.js.',
+    label: 'flowchart',
+    icon: flowchartIcon
+  }, {
+    title: 'Sequence Diagram',
+    subTitle: 'Render sequence diagram by js-sequence.',
+    label: 'sequence',
+    icon: sequenceIcon
+  }, {
+    title: 'Mermaid',
+    subTitle: 'Render Diagram by mermaid.',
+    label: 'mermaid',
+    icon: mermaidIcon
   }]
 }