+++ src/muya/lib/utils/exportMarkdown.js
@@ -70,7 +70,7 @@ class ExportMarkdown {
           this.insertLineBreak(result, indent)
           switch (block.functionType) {
             case 'table': {
-              const table = block.children[1]
+              const table = block.children[0]
               result.push(this.normalizeTable(table, indent))
               break
             }