+++ .github/CHANGELOG.md
@@ -43,6 +43,7 @@ foo<section>bar</section>zar
 - Support input inline Ruby charactors as raw html (#257)
 - Added unsaved tab indicator
 - Add front Menu by click the front menu icon (#875)
+- Support diagram: [flowchart](https://github.com/adrai/flowchart.js), [vega-lite](https://github.com/vega/vega-lite), [mermaid](https://github.com/knsv/mermaid), [sequence](https://github.com/bramp/js-sequence-diagrams) (#914)
 - Support create indent code block in preview mode.(#920)
 
 **:butterfly:Optimization**